optionally lock to specific Polar strap; honor "record heartrate" setting; deal with 64-bit issues
"Run" now compiles in release.
Fixed 64-bit compile warnings, but disabled arm64 due to using old testflight library and not wanting to upgrade and test just now.

+++ Shared/ChannelList.cpp
@@ -11,7 +11,7 @@
 // Not thread-safe;  caller must hold a lock on "mutex"
 
 int ChannelList::createList(const char *lst) {
-    int idx = lists.size();
+    int idx = (int)lists.size();
     
     ids[lst] = idx;
